+++ src/main/java/com/threestar/trainus/global/utils/PageLimitCalculator.java
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class PageLimitCalculator {
 
-	public static Long calculatePageLimit(Long page, Long pageSize, Long movablePageCount) {
+	public static int calculatePageLimit(int page, int pageSize, int movablePageCount) {
 		return (((page - 1) / movablePageCount) + 1) * pageSize * movablePageCount + 1;
 	}
 }
